moved newly created authentication manager to a new file
